server/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
greenlet==1.0.0
idna==2.10
IMDbPY==2020.9.25
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-dotenv==0.16.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.3
urllib3==1.26.4
Werkzeug==1.0.1
